Replace InjectValidatorBundle with Jersey's Feature added by default
