--- dropwizard-e2e/src/main/java/com/example/validation/InjectValidatorApp.java
@@ -1,20 +1,0 @@
-package com.example.validation;
-
-import io.dropwizard.Application;
-import io.dropwizard.Configuration;
-import io.dropwizard.setup.Bootstrap;
-import io.dropwizard.setup.Environment;
-import io.dropwizard.validation.InjectValidatorBundle;
-
-public class InjectValidatorApp extends Application<Configuration> {
-
-    @Override
-    public void initialize(Bootstrap<Configuration> bootstrap) {
-        bootstrap.addBundle(new InjectValidatorBundle());
-    }
-
-    @Override
-    public void run(Configuration configuration, Environment environment) throws Exception {
-        environment.jersey().register(ValidatedResource.class);
-    }
-}
